refactor(polyfills-loader): migrate tests from mocha/chai to node:test
Replace mocha globals and chai assertions with node:test and
node:assert/strict across 3 test files.

- `import type { PolyfillsLoaderConfig, PolyfillFile }` (type-only imports)
- `../src/*.js` -> `../dist/*.js`
- `expect(x).to.eql(y)` -> `assert.deepEqual(x, y)`
- `expect(x).to.equal(y)` -> `assert.equal(x, y)`
- `expect(x).to.be.a('string', msg)` -> `assert.equal(typeof x, 'string', msg)`
- `this.timeout(5000)` -> `{ timeout: 5000 }` option on describe
- `__dirname` -> `import.meta.dirname`
- Removed `test:update-snapshots` script
- Add `--experimental-strip-types` to override CI NODE_OPTIONS workaround
