+++ packages/polyfills-loader/test/createPolyfillsLoader.test.ts
@@ -1,9 +1,10 @@
-import { expect } from 'chai';
+import { describe, it } from 'node:test';
+import assert from 'node:assert/strict';
 import fs from 'fs';
 import path from 'path';
-import { PolyfillsLoaderConfig } from '../src/types.js';
-import { createPolyfillsLoader } from '../src/createPolyfillsLoader.js';
-import { noModuleSupportTest, fileTypes } from '../src/utils.js';
+import type { PolyfillsLoaderConfig } from '../dist/types.js';
+import { createPolyfillsLoader } from '../dist/createPolyfillsLoader.js';
+import { noModuleSupportTest, fileTypes } from '../dist/utils.js';
 
 const updateSnapshots = process.argv.includes('--update-snapshots');
 
@@ -14,26 +15,31 @@ interface TestSnapshotArgs {
 }
 
 async function testSnapshot({ name, config, expectedFiles = [] }: TestSnapshotArgs) {
-  const snapshotPath = path.join(__dirname, 'snapshots', 'createPolyfillsLoader', `${name}.js`);
+  const snapshotPath = path.join(
+    import.meta.dirname,
+    'snapshots',
+    'createPolyfillsLoader',
+    `${name}.js`,
+  );
   const loader = await createPolyfillsLoader(config);
   if (!loader) {
     throw new Error('No loader was generated');
   }
 
-  expect(loader.polyfillFiles.map(f => f.path)).to.eql(expectedFiles);
+  assert.deepEqual(
+    loader.polyfillFiles.map(f => f.path),
+    expectedFiles,
+  );
 
   if (updateSnapshots) {
     fs.writeFileSync(snapshotPath, loader.code, 'utf-8');
   } else {
     const snapshot = fs.readFileSync(snapshotPath, 'utf-8');
-    expect(loader.code.trim()).to.equal(snapshot.trim());
+    assert.equal(loader.code.trim(), snapshot.trim());
   }
 }
 
-describe('createPolyfillsLoader', function describe() {
-  // bootup of the first test can take a long time in CI to load all the polyfills
-  this.timeout(5000);
-
+describe('createPolyfillsLoader', { timeout: 5000 }, () => {
   it('generates a loader script with one module resource', async () => {
     await testSnapshot({
       name: 'module-resource',
